front: switchSNCF: typescript
