+++ front/src/common/BootstrapSNCF/SwitchSNCF/SwitchSNCF.scss
